+++ test/karma/formatters.spec.js
@@ -14,7 +14,7 @@ import * as formatters from '../../src/formatters'
 group('Vanilla Formatter', () => {
   test('add new error', (assert) => {
     const formatter = new formatters.Vanilla()
-    formatter.addError('validation failed', 'username', 'required')
+    formatter.addError('validation failed', 'username', 'required', [])
     assert.deepEqual(formatter.toJSON(), [
       {
         field: 'username',
@@ -28,13 +28,15 @@ group('Vanilla Formatter', () => {
 group('JsonApi Formatter', () => {
   test('add new error', (assert) => {
     const formatter = new formatters.JsonApi()
-    formatter.addError('validation failed', 'username', 'required')
-    assert.deepEqual(formatter.toJSON(), [
-      {
-        source: { pointer: 'username' },
-        detail: 'validation failed',
-        title: 'required'
-      }
-    ])
+    formatter.addError('validation failed', 'username', 'required', [])
+    assert.deepEqual(formatter.toJSON(), {
+      errors: [
+        {
+          source: { pointer: 'username' },
+          detail: 'validation failed',
+          title: 'required'
+        }
+      ]
+    })
   })
 })