+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+<a name="5.0.2"></a>
+## [5.0.2](https://github.com/poppinss/indicative/compare/v5.0.1...v5.0.2) (2018-01-28)
+
+
+
+<a name="5.0.1"></a>
+## [5.0.1](https://github.com/poppinss/indicative/compare/v5.0.0...v5.0.1) (2018-01-28)
+
+
+### Bug Fixes
+
+* **validator:** checks for errors length when it's an array ([5119ea9](https://github.com/poppinss/indicative/commit/5119ea9))
+
+
+
+<a name="5.0.0"></a>
+# [5.0.0](https://github.com/poppinss/indicative/compare/v4.0.4...v5.0.0) (2018-01-28)
+
+
+### Bug Fixes
+
+* **formatters:** fix jsonapi formatter to return errors in right format ([eb74497](https://github.com/poppinss/indicative/commit/eb74497))
+
+
+### Features
+
+* **formatter:** now receives additional args argument ([8826d27](https://github.com/poppinss/indicative/commit/8826d27))
+
+
+### BREAKING CHANGES
+
+* **formatters:** Instead of returning an array, an object with key `error` is returned containing an
+array of errors
+
+
+
 <a name="4.0.4"></a>
 ## [4.0.4](https://github.com/poppinss/indicative/compare/v4.0.3...v4.0.4) (2018-01-12)